Combine minor modes into one
Override mode is now just a variable that controls whether more keys are
added to the keymap.

Besides simplifying the package, this fixes a problem where it was not
possible to disable the override mode in a local buffer without
disabling the global version of the mode. Quoting the docstring of
`define-globalized-minor-mode`, this is what should happen

"When a major mode is initialized, MODE is actually turned on just
after running the major mode's hook.  However, MODE is not turned
on if the hook has explicitly disabled it."

It seems that the problem was caused by "chaining" the two modes
together.